Using trading view socket

+++ src/strategies/main/types.go
@@ -2,10 +2,10 @@ package mainstrategy
 
 // Candle ...
 type Candle struct {
-	Open      float32
-	High      float32
-	Low       float32
-	Close     float32
+	Open      float64
+	High      float64
+	Low       float64
+	Close     float64
 	Volume    float64
 	Timestamp int64
 }
